Refactor tests with pytest
