+++ wfc/wfc_adjacency.py
@@ -47,40 +47,3 @@ def adjacency_extraction(
                 if is_valid_overlap_xy(direction, pattern_1, pattern_2):
                     legal.append((direction, pattern_1, pattern_2))
     return legal
-
-
-def test_adjacency_extraction():
-    from wfc_tiles import make_tile_catalog
-    from wfc_patterns import make_pattern_catalog
-    import imageio
-
-    # TODO: generalize this to more than the four cardinal directions
-    direction_offsets = list(enumerate([(0, -1), (1, 0), (0, 1), (-1, 0)]))
-
-    filename = "../images/samples/Red Maze.png"
-    img = imageio.imread(filename)
-    tile_size = 1
-    pattern_width = 2
-    rotations = 0
-    tile_catalog, tile_grid, code_list, unique_tiles = make_tile_catalog(img, tile_size)
-    pattern_catalog, pattern_weights, pattern_list, pattern_grid = make_pattern_catalog(
-        tile_grid, pattern_width, rotations
-    )
-    adjacency_relations = adjacency_extraction(
-        pattern_grid, pattern_catalog, direction_offsets
-    )
-    assert ((0, -1), -6150964001204120324, -4042134092912931260) in adjacency_relations
-    assert ((-1, 0), -4042134092912931260, 3069048847358774683) in adjacency_relations
-    assert ((1, 0), -3950451988873469076, -3950451988873469076) in adjacency_relations
-    assert ((-1, 0), -3950451988873469076, -3950451988873469076) in adjacency_relations
-    assert ((0, 1), -3950451988873469076, 3336256675067683735) in adjacency_relations
-    assert (
-        not ((0, -1), -3950451988873469076, -3950451988873469076) in adjacency_relations
-    )
-    assert (
-        not ((0, 1), -3950451988873469076, -3950451988873469076) in adjacency_relations
-    )
-
-
-if __name__ == "__main__":
-    test_adjacency_extraction()
