+++ wfc/wfc_solver.py
@@ -441,235 +441,3 @@ def run(
             if onFinal:
                 onFinal(wave)
             raise
-
-
-#############################
-# Tests
-
-
-def test_makeWave():
-    wave = makeWave(3, 10, 20, ground=[-1])
-    # print(wave)
-    # print(wave.sum())
-    # print((2*10*19) + (1*10*1))
-    assert wave.sum() == (2 * 10 * 19) + (1 * 10 * 1)
-    assert wave[2, 5, 19] == True
-    assert wave[1, 5, 19] == False
-
-
-def test_entropyLocationHeuristic():
-    wave = numpy.ones((5, 3, 4), dtype=bool)  # everthing is possible
-    wave[1:, 0, 0] = False  # first cell is fully observed
-    wave[4, :, 2] = False
-    preferences = numpy.ones((3, 4), dtype=float) * 0.5
-    preferences[1, 2] = 0.3
-    preferences[1, 1] = 0.1
-    heu = makeEntropyLocationHeuristic(preferences)
-    result = heu(wave)
-    assert [1, 2] == result
-
-
-def test_observe():
-
-    my_wave = numpy.ones((5, 3, 4), dtype=bool)
-    my_wave[0, 1, 2] = False
-
-    def locHeu(wave):
-        assert numpy.array_equal(wave, my_wave)
-        return 1, 2
-
-    def patHeu(weights):
-        assert numpy.array_equal(weights, my_wave[:, 1, 2])
-        return 3
-
-    assert observe(my_wave, locationHeuristic=locHeu, patternHeuristic=patHeu) == (
-        3,
-        1,
-        2,
-    )
-
-
-def test_propagate():
-    wave = numpy.ones((3, 3, 4), dtype=bool)
-    adjLists = {}
-    # checkerboard #0/#1 or solid fill #2
-    adjLists[(+1, 0)] = adjLists[(-1, 0)] = adjLists[(0, +1)] = adjLists[(0, -1)] = [
-        [1],
-        [0],
-        [2],
-    ]
-    wave[:, 0, 0] = False
-    wave[0, 0, 0] = True
-    adj = makeAdj(adjLists)
-    propagate(wave, adj, periodic=False)
-    expected_result = numpy.array(
-        [
-            [
-                [True, False, True, False],
-                [False, True, False, True],
-                [True, False, True, False],
-            ],
-            [
-                [False, True, False, True],
-                [True, False, True, False],
-                [False, True, False, True],
-            ],
-            [
-                [False, False, False, False],
-                [False, False, False, False],
-                [False, False, False, False],
-            ],
-        ]
-    )
-    assert numpy.array_equal(wave, expected_result)
-
-
-def test_run():
-    wave = makeWave(3, 3, 4)
-    adjLists = {}
-    adjLists[(+1, 0)] = adjLists[(-1, 0)] = adjLists[(0, +1)] = adjLists[(0, -1)] = [
-        [1],
-        [0],
-        [2],
-    ]
-    adj = makeAdj(adjLists)
-
-    first_result = run(
-        wave.copy(),
-        adj,
-        locationHeuristic=lexicalLocationHeuristic,
-        patternHeuristic=lexicalPatternHeuristic,
-        periodic=False,
-    )
-
-    expected_first_result = numpy.array([[0, 1, 0, 1], [1, 0, 1, 0], [0, 1, 0, 1]])
-
-    assert numpy.array_equal(first_result, expected_first_result)
-
-    event_log = []
-
-    def onChoice(pattern, i, j):
-        event_log.append((pattern, i, j))
-
-    def onBacktrack():
-        event_log.append("backtrack")
-
-    second_result = run(
-        wave.copy(),
-        adj,
-        locationHeuristic=lexicalLocationHeuristic,
-        patternHeuristic=lexicalPatternHeuristic,
-        periodic=True,
-        backtracking=True,
-        onChoice=onChoice,
-        onBacktrack=onBacktrack,
-    )
-
-    expected_second_result = numpy.array([[2, 2, 2, 2], [2, 2, 2, 2], [2, 2, 2, 2]])
-
-    assert numpy.array_equal(second_result, expected_second_result)
-    print(event_log)
-    assert event_log == [(0, 0, 0), "backtrack"]
-
-    def explode(wave):
-        if wave.sum() < 20:
-            raise Infeasible
-
-    try:
-        result = run(
-            wave.copy(),
-            adj,
-            locationHeuristic=lexicalLocationHeuristic,
-            patternHeuristic=lexicalPatternHeuristic,
-            periodic=True,
-            backtracking=True,
-            checkFeasible=explode,
-        )
-        print(result)
-        happy = False
-    except Contradiction:
-        happy = True
-
-    assert happy
-
-
-def test_recurse_vs_loop():
-    from wfc_tiles import make_tile_catalog
-    from wfc_patterns import make_pattern_catalog, pattern_grid_to_tiles
-    from wfc_adjacency import adjacency_extraction
-    from wfc_solver import (
-        run,
-        makeWave,
-        makeAdj,
-        lexicalLocationHeuristic,
-        lexicalPatternHeuristic,
-    )
-    from wfc_visualize import (
-        figure_list_of_tiles,
-        figure_false_color_tile_grid,
-        figure_pattern_catalog,
-        render_tiles_to_output,
-        figure_adjacencies,
-    )
-
-    import imageio
-
-    img = imageio.imread("../images/samples/Red Maze.png")
-    tile_size = 1
-    pattern_width = 2
-    rotations = 0
-    output_size = [84, 84]
-    ground = None
-    direction_offsets = list(enumerate([(0, -1), (1, 0), (0, 1), (-1, 0)]))
-    tile_catalog, tile_grid, code_list, unique_tiles = make_tile_catalog(img, tile_size)
-    pattern_catalog, pattern_weights, pattern_list, pattern_grid = make_pattern_catalog(
-        tile_grid, pattern_width, rotations
-    )
-    adjacency_relations = adjacency_extraction(
-        pattern_grid, pattern_catalog, direction_offsets
-    )
-    number_of_patterns = len(pattern_weights)
-    encode_patterns = dict(enumerate(pattern_list))
-    decode_patterns = {x: i for i, x in enumerate(pattern_list)}
-    decode_directions = {j: i for i, j in direction_offsets}
-    adjacency_list = {}
-    for i, d in direction_offsets:
-        adjacency_list[d] = [set() for i in pattern_weights]
-    for i in adjacency_relations:
-        adjacency_list[i[0]][decode_patterns[i[1]]].add(decode_patterns[i[2]])
-    wave = makeWave(number_of_patterns, output_size[0], output_size[1])
-    adjacency_matrix = makeAdj(adjacency_list)
-    solution_loop = run(
-        wave.copy(),
-        adjacency_matrix,
-        locationHeuristic=lexicalLocationHeuristic,
-        patternHeuristic=lexicalPatternHeuristic,
-        periodic=True,
-        backtracking=False,
-        onChoice=None,
-        onBacktrack=None,
-    )
-    solution_recurse = run_recurse(
-        wave.copy(),
-        adjacency_matrix,
-        locationHeuristic=lexicalLocationHeuristic,
-        patternHeuristic=lexicalPatternHeuristic,
-        periodic=True,
-        backtracking=False,
-        onChoice=None,
-        onBacktrack=None,
-    )
-    assert numpy.array_equiv(solution_loop, solution_recurse)
-
-
-from pycallgraph import PyCallGraph
-from pycallgraph.output import GraphvizOutput
-
-if __name__ == "__main__":
-    with PyCallGraph(output=GraphvizOutput()):
-        test_makeWave()
-        test_entropyLocationHeuristic()
-        test_observe()
-        test_propagate()
-        test_run()
-        # test_recurse_vs_loop()
